+++ CHANGELOG.md
@@ -1,13 +1,17 @@
 CHANGELOG
 =========
 
+version 1.0.3
+-------------
+* Patch: Fix how the `Authorization` header is built.
+
 version 1.0.2
 -------------
-* Minor: Change information about the plugin
+* Patch: Change information about the plugin
 
 version 1.0.1
 -------------
-* Minor: Update jenkins tests
+* Patch: Update jenkins tests
 
 version 1.0.0
 -------------